setup/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
voyageai>=0.2.0
openai>=1.0.0
pymongo>=4.6.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
PyMuPDF>=1.23.0
